Finished Integration Tests
