+++ TSQLLINT_TESTS/TSQLLINT_LIB_TESTS.csproj
@@ -46,7 +46,8 @@
   <ItemGroup>
     <Compile Include="Helpers\HelperTests.cs" />
     <Compile Include="Helpers\TestHelper.cs" />
-    <Compile Include="IntegrationTests\IntegrationTests.cs" />
+    <Compile Include="IntegrationTests\IntegrationMissingConfigTests.cs" />
+    <Compile Include="IntegrationTests\IntegrationExistingConfigTests.cs" />
     <Compile Include="UnitTests\CommandLineOptions\CommandLineOptionHandlerTest.cs" />
     <Compile Include="UnitTests\CommandLineOptions\CommandLineParserTest.cs" />
     <Compile Include="UnitTests\Config\ConfigFileGeneratorTests.cs" />
